added a confirm dialogue to deletes and adjusted css

+++ src/components/DeleteArticle.jsx
@@ -15,35 +15,41 @@ export default function DeleteArticle({
   const [error, setError] = useState(null);
   const [isDeleting, setIsDeleting] = useState(false);
 
-  function handleDelete() {
-    setIsDeleting(true);
-    if (articles && setArticles) {
-      const originalArticles = [...articles];
-      const optimisticDelete = articles.filter((article) => {
-        return article.article_id != id;
-      });
-      setArticles(optimisticDelete);
+  function handleDelete(e) {
+    e.preventDefault();
+
+    if (window.confirm("Are you sure you want to delete this article?")) {
+      setIsDeleting(true);
+      if (articles && setArticles) {
+        const originalArticles = [...articles];
+        const optimisticDelete = articles.filter((article) => {
+          return article.article_id != id;
+        });
+        setArticles(optimisticDelete);
+      }
+      deleteArticle(id)
+        .then(() => {
+          if (!articles) {
+            nav("/");
+          }
+        })
+        .catch((err) => {
+          setError("Failed to delete article");
+          if (setArticles) {
+            setArticles(originalArticles);
+          }
+        })
+        .finally(() => {
+          setIsDeleting(false);
+        });
+    } else {
+      //cancel
     }
-    deleteArticle(id)
-      .then(() => {
-        if (!articles) {
-          nav("/");
-        }
-      })
-      .catch((err) => {
-        setError("Failed to delete article");
-        if (setArticles) {
-          setArticles(originalArticles);
-        }
-      })
-      .finally(() => {
-        setIsDeleting(false);
-      });
   }
 
   return (
     <div className="delete-div">
-      <button className="delete-button" onClick={handleDelete}>
+      <button className="delete-button" onClick={(e) => handleDelete(e)}>
         <Bin />
       </button>
       {error ? <p className="delete-error">{error}</p> : null}
